gpu: Remove gpu-av path from corechecks
