+++ build-android/cmake/layerlib/CMakeLists.txt
@@ -68,7 +68,6 @@ add_library(VkLayer_khronos_validation SHARED
         ${SRC_DIR}/layers/buffer_validation.cpp
         ${SRC_DIR}/layers/shader_validation.cpp
         ${SRC_DIR}/layers/gpu_validation.cpp
-        ${SRC_DIR}/layers/gpuav_validation.cpp
         ${SRC_DIR}/layers/best_practices.cpp
         ${COMMON_DIR}/include/layer_chassis_dispatch.cpp
         ${COMMON_DIR}/include/chassis.cpp
